ズームレベル検索結果が、地図タイルURLの変化時に無効になるようにした。リファクタした。
