+++ Editor/CityImport/CityLoadConfigGUI.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using PLATEAU.CityImport.Setting;
+using PLATEAU.Editor.CityImport.PackageLodSettingGUIs;
 using PLATEAU.Editor.EditorWindow.Common;
-using PLATEAU.Dataset;
-using PLATEAU.PolygonMesh;
-using PLATEAU.Util;
 using UnityEditor;
 using UnityEngine;
